hotelesbackend/requirements.txt
asgiref==3.2.3
Django==3.0.3
django-filter==2.2.0
django-polymorphic==2.1.2
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
